modules/microservices/location-producer/requirements.txt
grpcio-tools==1.58.0
kafka-python==2.0.2